Turns the outcome of runExperiment into a long table.

Takes the output of runExperiment, extracts the dataframes containing the ecological memory patterns generated by computeMemory, and binds them together into a single dataframe ready for further analyses or plotting.

experimentToTable(
 experiment.output = NULL,
 parameters.file = NULL,
 sampling.names = NULL,
 R2 = TRUE
 )

Arguments

experiment.output
list, output of runExperiment.

parameters.file
dataframe of simulation parameters.

sampling.names
vector of character strings with the names of the columns of simulations.file.

R2
boolean. If TRUE, the average pseudo R-squared of the random forest models used to analyze the ecological memory pattern of the virtual taxa are shown with the taxon traits.

Value

A dataframe.

Details

This function is used internally by plotExperiment, but it is also available to users in case they want to do other kinds of analyses or plots with the data.

See also