Lagged data generated by prepareLaggedData.

A dataframe resulting from the application of prepareLaggedData to the dataset palaeodata. The dataframe columns are:

data(palaeodataLagged)

Format

dataframe with 13 columns and 3988 rows.

Details

Response_0 numeric, values of the response variable selected by the user in the argument response of the function prepareLaggedData. This column is used as response variable by the function computeMemory. In this case, Response represent pollen counts of Pinus.

Response_0.2-1 numeric, time delayed values of the response for different lags (in ky). Considered together these columns represent the endogenous ecological memory.

climate.temperatureAverage_0 numeric, values of the variable temperatureAverage for the lag 0 (no lag). This column represents the concurrent effect of the temperature over the response. #'

climate.rainfallAverage_0 numeric, values of the variable rainfallAverage for the lag 0 (no lag). This column represents the concurrent effect of rainfall over the response.

climate.temperatureAverage_0.2-1 numeric, time delayed values of temperatureAverage for lags 0.2 to 1 (in ky). #'

climate.rainfallAverage_0.2-1 numeric, time delayed values of rainfallAverage for lags 0.2 to 1 (in ky).