Output of computeMemory

List containing the output of computeMemory applied to palaeodataLagged. Its slots are:

data(palaeodataMemory)

Format

List with four slots.

Details

memory dataframe with five columns:

Variable character, names and lags of the different variables used to model ecological memory.

median numeric, median importance across repetitions of the given Variable according to Random Forest.

sd numeric, standard deviation of the importance values of the given Variable across repetitions.

min and max numeric, percentiles 0.05 and 0.95 of importance values of the given Variable across repetitions.

R2 vector, values of pseudo R-squared value obtained for the Random Forest model fitted on each repetition. Pseudo R-squared is the Pearson correlation beteween the observed and predicted data.

prediction dataframe, with the same columns as the dataframe in the slot memory, with the median and confidence intervals of the predictions of all random forest models fitted.

multicollinearity multicollinearity analysis on the input data performed with vif. A vif value higher than 5 indicates that the given variable is highly correlated with other variables.