Plots response surfaces for tree-based models.

Plots a response surface plot or interaction plot (2 predictors and a model response) for models of the functions ranger, randomForest, and rpart. It also plots the observed data on top of the predicted surface.

Usage

plotInteraction(
  model = NULL,
  data = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  grid = 100,
  point.size.range = c(0.1, 1)
  )

Arguments

model

a model object produced by the functions ranger, randomForest, or rpart.

data

dataframe used to fit the model.

x

character string, name of column in data to be plotted in the x axis.

y

character string, name of column in data to be plotted in the y axis.

z

character string, name of the response variable column in data, plotted as the surface.

grid

numeric, resolution of the x and y axes.

point.size.range

numeric vector with two values defining the range size of the points representing the observed data.

Value

A ggplot object.

Author

Blas M. Benito <blasbenito@gmail.com>