Output of computeMemory

List containing the output of computeMemory applied to palaeodataLagged. Its slots are:

Usage

data(palaeodataMemory)

Format

List with five slots.

Details

response character, response variable name.

drivers character vector, driver variable names.

memory dataframe with five columns:

variable character, names of the different variables used to model ecological memory.

lag numeric, time lag values.

median numeric, median importance across repetitions of the given variable according to Random Forest.

sd numeric, standard deviation of the importance values of the given variable across repetitions.

min and max numeric, percentiles 0.05 and 0.95 of importance values of the given variable across repetitions.

R2 vector, values of pseudo R-squared value obtained for the Random Forest model fitted on each repetition. Pseudo R-squared is the Pearson correlation between the observed and predicted data.

prediction dataframe, with the same columns as the dataframe in the slot memory, with the median and confidence intervals of the predictions of all random forest models fitted.

See also

Author

Blas M. Benito <blasbenito@gmail.com>